requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
paramiko==3.4.0
pywinrm==0.4.3
requests-ntlm==1.2.0

